fix problem on rich text editor
